fixes for unsigned

+++ main.h
@@ -5,7 +5,7 @@ int _printf(const char *format, ...);
 int _putchar(char c);
 void reverse(char s[], int length);
 char *convert(int n, char *s, int base);
-char *convertu(long int n, char *s, int base);
+char *convertu(unsigned int n, char *s, int base);
 
 
 #endif /* _main_H_ */
